third_party/blink/web_tests/images/style-access-during-imageChanged-crash-expected.txt

CONSOLE WARNING: Failed to decode downloaded font: data:text/plain,1
CONSOLE WARNING: OTS parsing error: file less than 4 bytes
This test passes if it does not cause an assertion failure or a crash.

A 
